+++ tests/core/test_playback.py
@@ -6,7 +6,7 @@ import pytest
 from mopidy import backend, core
 from mopidy.internal import deprecation
 from mopidy.internal.models import PlaybackState
-from mopidy.models import Track
+from mopidy.models import TlTrack, Track
 
 from tests import dummy_audio, dummy_backend
 
@@ -77,8 +77,7 @@ class BaseTest:
         # We don't have a core actor running, so call about to finish directly.
         self.audio.set_about_to_finish_callback(self.playback._on_about_to_finish)
 
-        with deprecation.ignore():
-            self.core.tracklist.add(self.tracks)
+        self.add_tracks(self.tracks)
 
         self.events = []
         self.patcher = mock.patch("mopidy.audio.listener.AudioListener.send")
@@ -92,6 +91,14 @@ class BaseTest:
     def teardown_method(self):
         pykka.ActorRegistry.stop_all()
         self.patcher.stop()
+
+    def add_tracks(
+        self,
+        tracks: list[Track],
+        at_position: int | None = None,
+    ) -> list[TlTrack]:
+        self.backend.library.dummy_library = tracks
+        return self.core.tracklist.add([t.uri for t in tracks], at_position=at_position)
 
     def replay_events(self, until=None):
         while self.events:
@@ -251,8 +258,7 @@ class TestNextHandling(BaseTest):
     def test_next_skips_over_change_track_error(self):
         # Trigger an exception in translate_uri.
         track = Track(uri="dummy:error", length=1234)
-        with deprecation.ignore():
-            self.core.tracklist.add(tracks=[track], at_position=1)
+        self.add_tracks([track], at_position=1)
 
         tl_tracks = self.core.tracklist.get_tl_tracks()
 
@@ -267,8 +273,7 @@ class TestNextHandling(BaseTest):
     def test_next_skips_over_change_track_unplayable(self):
         # Make translate_uri return None.
         track = Track(uri="dummy:unplayable", length=1234)
-        with deprecation.ignore():
-            self.core.tracklist.add(tracks=[track], at_position=1)
+        self.add_tracks([track], at_position=1)
 
         tl_tracks = self.core.tracklist.get_tl_tracks()
 
@@ -361,8 +366,7 @@ class TestPreviousHandling(BaseTest):
     def test_previous_skips_over_change_track_error(self):
         # Trigger an exception in translate_uri.
         track = Track(uri="dummy:error", length=1234)
-        with deprecation.ignore():
-            self.core.tracklist.add(tracks=[track], at_position=1)
+        self.add_tracks([track], at_position=1)
 
         tl_tracks = self.core.tracklist.get_tl_tracks()
 
@@ -377,8 +381,7 @@ class TestPreviousHandling(BaseTest):
     def test_previous_skips_over_change_track_unplayable(self):
         # Makes translate_uri return None.
         track = Track(uri="dummy:unplayable", length=1234)
-        with deprecation.ignore():
-            self.core.tracklist.add(tracks=[track], at_position=1)
+        self.add_tracks([track], at_position=1)
 
         tl_tracks = self.core.tracklist.get_tl_tracks()
 
@@ -403,8 +406,7 @@ class TestOnAboutToFinish(BaseTest):
     def test_on_about_to_finish_skips_over_change_track_error(self):
         # Trigger an exception in translate_uri.
         track = Track(uri="dummy:error", length=1234)
-        with deprecation.ignore():
-            self.core.tracklist.add(tracks=[track], at_position=1)
+        self.add_tracks([track], at_position=1)
 
         tl_tracks = self.core.tracklist.get_tl_tracks()
 
@@ -418,8 +420,7 @@ class TestOnAboutToFinish(BaseTest):
     def test_on_about_to_finish_skips_over_change_track_unplayable(self):
         # Makes translate_uri return None.
         track = Track(uri="dummy:unplayable", length=1234)
-        with deprecation.ignore():
-            self.core.tracklist.add(tracks=[track], at_position=1)
+        self.add_tracks([track], at_position=1)
 
         tl_tracks = self.core.tracklist.get_tl_tracks()
 
@@ -848,6 +849,7 @@ class TestUnplayableURI(BaseTest):
 
     def test_pause_changes_state_even_if_track_is_unplayable(self):
         self.core.playback.pause()
+
         assert self.core.playback.get_state() == core.PlaybackState.PAUSED
 
     def test_resume_does_nothing_if_track_is_unplayable(self):
@@ -886,9 +888,9 @@ class TestSeek(BaseTest):
 
     def test_seek_fails_for_track_without_duration(self):
         track = self.tracks[0].replace(length=None)
+        self.add_tracks([track])
         self.core.tracklist.clear()
-        with deprecation.ignore():
-            self.core.tracklist.add(tracks=[track])
+        self.core.tracklist.add(uris=[track.uri])
 
         self.core.playback.play()
         self.replay_events()
@@ -997,8 +999,8 @@ class TestStream(BaseTest):
 class TestBug1871Regression(BaseTest):
     def test(self):
         track = Track(uri="dummy:d", length=1234, name="baz")
-        with deprecation.ignore():
-            self.core.tracklist.add(tracks=[track], at_position=1)
+        self.add_tracks([track])
+        self.core.tracklist.add(uris=[track.uri], at_position=1)
 
         self.core.playback.play()
         self.replay_events()
@@ -1014,7 +1016,7 @@ class TestBug1871Regression(BaseTest):
         assert self.playback.get_stream_title() is None
 
 
-class TestBackendSelection:
+class TestBackendSelection(BaseTest):
     def setup_method(self, method):
         config = {"core": {"max_tracklist_length": 10000}}
 
@@ -1027,11 +1029,6 @@ class TestBackendSelection:
         self.backend2.uri_schemes.get.return_value = ["dummy2"]
         self.playback2 = mock.Mock(spec=backend.PlaybackProvider)
         self.backend2.playback = self.playback2
-
-        self.tracks = [
-            Track(uri="dummy1:a", length=40000),
-            Track(uri="dummy2:a", length=40000),
-        ]
 
         self.core = core.Core(
             config, mixer=None, backends=[self.backend1, self.backend2]
@@ -1183,18 +1180,19 @@ class TestBug1177Regression:
     def test(self):
         config = {"core": {"max_tracklist_length": 10000}}
 
+        track1 = Track(uri="dummy:a", length=40000)
+        track2 = Track(uri="dummy:b", length=40000)
+
         b = mock.Mock()
         b.uri_schemes.get.return_value = ["dummy"]
+        b.library = mock.Mock(spec=backend.LibraryProvider)
+        b.library.lookup.return_value.get.side_effect = [[track1], [track2]]
         b.playback = mock.Mock(spec=backend.PlaybackProvider)
         b.playback.change_track.return_value.get.return_value = True
         b.playback.play.return_value.get.return_value = True
 
-        track1 = Track(uri="dummy:a", length=40000)
-        track2 = Track(uri="dummy:b", length=40000)
-
         c = core.Core(config, mixer=None, backends=[b])
-        with deprecation.ignore():
-            c.tracklist.add(tracks=[track1, track2])
+        c.tracklist.add([track1.uri, track2.uri])
 
         c.playback.play()
         b.playback.change_track.assert_called_once_with(track1)
@@ -1296,8 +1294,7 @@ class TestEndlessLoop(BaseTest):
 
     def test_play(self):
         self.core.tracklist.clear()
-        with deprecation.ignore():
-            self.core.tracklist.add(tracks=self.tracks_play)
+        self.add_tracks(self.tracks_play)
 
         self.backend.playback.reset_call_limit().get()
         self.core.tracklist.set_repeat(True)
@@ -1310,8 +1307,7 @@ class TestEndlessLoop(BaseTest):
 
     def test_next(self):
         self.core.tracklist.clear()
-        with deprecation.ignore():
-            self.core.tracklist.add(tracks=self.tracks_other)
+        self.add_tracks(self.tracks_other)
 
         self.backend.playback.reset_call_limit().get()
         self.core.tracklist.set_repeat(True)
@@ -1327,8 +1323,7 @@ class TestEndlessLoop(BaseTest):
 
     def test_previous(self):
         self.core.tracklist.clear()
-        with deprecation.ignore():
-            self.core.tracklist.add(tracks=self.tracks_other)
+        self.add_tracks(self.tracks_other)
 
         self.backend.playback.reset_call_limit().get()
         self.core.tracklist.set_repeat(True)
@@ -1344,8 +1339,7 @@ class TestEndlessLoop(BaseTest):
 
     def test_on_about_to_finish(self):
         self.core.tracklist.clear()
-        with deprecation.ignore():
-            self.core.tracklist.add(tracks=self.tracks_other)
+        self.add_tracks(self.tracks_other)
 
         self.backend.playback.reset_call_limit().get()
         self.core.tracklist.set_repeat(True)